twister: Add robot framework integration
This change adds an entrypoint for the common robot framework and a
first library to interact with the board, similar to the functionality
currently available in pytest.

To reuse the command line parsing and have a proper twister config, we
provide a robot_wrapper which serializes the configuration and reloads
it in the library.

This change addresses #64825, which asks for robot scripts without
renode.

Signed-off-by: Stefan Kraus <stefan.kraus@ul.com>